+++ ChangeLog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 
+## Version 1.4
+
+- Add workstations filter on fullcalendar tasks view
+
+## Version 1.3
+
+- Add getNomUrl() workstation linked to task on fullCalendar tasks view
 
 ## Version 1.2